Fix in-container detection for Docker
Previously, the InDockerContainer was checking whether the control
group of the Linux process contains "docker". This was working well
years ago, but apparently not any longer.

We add a few alternatives to check so that we also check for
"/.dockerenv" when the Docker Engine used, or checking for "lxc"
in the process environment for LXC.

Tests were not extended.